change get int to get float

+++ Platformer 2d URP Sandbox/Assets/Scripts/V Script/UI/HighScore.cs
@@ -12,7 +12,7 @@ public class HighScore : MonoBehaviour
 
     private void Start()
     {
-        highscore = PlayerPrefs.GetInt("HighScore", 0);
+        highscore = PlayerPrefs.GetFloat("HighScore", 0);
     }
     private void Update()
     {
